requirements.txt
azure-storage-blob==12.14.1
runpod==1.7.7
torch==2.5.1
xformers==0.0.28.post3
transformers==4.47.1
accelerate>=0.34.1
bitsandbytes>=0.43.3
trl==0.11.4
peft==0.14.0
triton==3.1.0
PyPDF2==2.11.2
PyMuPDF==1.25.1
pytesseract==0.3.10
Pillow==10.4.0
openai>=1.0.0
datasets>=2.13.1,<3.0.0
packaging==23.0
tqdm>=4.66.3
unsloth==2025.1.6
vllm==0.6.6.post1
sentence-transformers==3.3.1
rouge-score==0.1.2
nltk==3.9.1
requests==2.32.3
gradio==5.9.1
psutil==6.1.1
flask==3.1.0
autoawq==0.2.7.post3
tensorboard==2.18.0
pdfplumber==0.11.5
gmft==0.4.0